+++ debezium-connector-mysql/src/test/java/io/debezium/connector/mysql/BinlogReaderIT.java
@@ -9,6 +9,7 @@ import static org.fest.assertions.Assertions.assertThat;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
 
+import io.debezium.connector.mysql.MySqlConnectorConfig.SecureConnectionMode;
 import java.nio.file.Path;
 import java.sql.Connection;
 import java.sql.SQLException;
@@ -50,6 +51,8 @@ public class BinlogReaderIT {
     private static final Path DB_HISTORY_PATH = Testing.Files.createTestingPath("file-db-history-binlog.txt").toAbsolutePath();
     private final UniqueDatabase DATABASE = new UniqueDatabase("logical_server_name", "connector_test_ro")
             .withDbHistoryPath(DB_HISTORY_PATH);
+
+    private static final String SET_TLS_PROTOCOLS = "database.enabledTLSProtocols";
 
     private Configuration config;
     private MySqlTaskContext context;
@@ -104,6 +107,18 @@ public class BinlogReaderIT {
             }
         }
         return counter.get();
+    }
+
+    protected long filterAtLeast(final int minNumber, final long timeout, final TimeUnit unit) throws InterruptedException {
+        final BinlogReaderMetrics metrics = reader.getMetrics();
+        final long initialFilterCount = metrics.getNumberOfEventsFiltered();
+        final long targetNumber = initialFilterCount + minNumber;
+        long startTime = System.currentTimeMillis();
+        while (metrics.getNumberOfEventsFiltered() < targetNumber && (System.currentTimeMillis() - startTime) < unit.toMillis(timeout)) {
+            // Ignore the records polled.
+            reader.poll();
+        }
+        return reader.getMetrics().getNumberOfEventsFiltered() - initialFilterCount;
     }
 
     protected Configuration.Builder simpleConfig() {
@@ -237,6 +252,46 @@ public class BinlogReaderIT {
         assertThat(orders.numberOfTombstones()).isEqualTo(0);
         assertThat(orders.numberOfKeySchemaChanges()).isEqualTo(1);
         assertThat(orders.numberOfValueSchemaChanges()).isEqualTo(1);
+    }
+
+    /**
+     * Setup a DATABASE_WHITELIST filter that filters all events.
+     * Verify all events are properly filtered.
+     * Verify numberOfFilteredEvents metric is incremented correctly.
+     */
+    @Test
+    @FixFor( "DBZ-1206" )
+    public void shouldFilterAllRecordsBasedOnDatabaseWhitelistFilter() throws Exception {
+        // Define configuration that will ignore all events from MySQL source.
+        config = simpleConfig()
+            .with(MySqlConnectorConfig.DATABASE_WHITELIST, "db-does-not-exist")
+            .build();
+
+        final Filters filters = new Filters.Builder(config).build();
+        context = new MySqlTaskContext(config, filters);
+        context.start();
+        context.source().setBinlogStartPoint("", 0L); // start from beginning
+        context.initializeHistory();
+        reader = new BinlogReader("binlog", context, new AcceptAllPredicate());
+
+        // Start reading the binlog ...
+        reader.start();
+
+        // Lets wait for at least 35 events to be filtered.
+        final int expectedFilterCount = 35;
+        final long numberFiltered = filterAtLeast(expectedFilterCount, 20, TimeUnit.SECONDS);
+
+        // All events should have been filtered.
+        assertThat(numberFiltered).isGreaterThanOrEqualTo(expectedFilterCount);
+
+        // There should be no schema changes
+        assertThat(schemaChanges.recordCount()).isEqualTo(0);
+
+        // There should be no records
+        assertThat(store.collectionCount()).isEqualTo(0);
+
+        // There should be no skipped
+        assertThat(reader.getMetrics().getNumberOfSkippedEvents()).isEqualTo(0);
     }
 
     @Test
@@ -385,6 +440,52 @@ public class BinlogReaderIT {
         inconsistentSchema(EventProcessingFailureHandlingMode.IGNORE);
         int consumed = consumeAtLeast(2, 2, TimeUnit.SECONDS);
         assertThat(consumed).isZero();
+    }
+
+    @Test(expected = ConnectException.class)
+    @FixFor( "DBZ-1208" )
+    public void shouldFailOnUnknownTlsProtocol() {
+        final UniqueDatabase REGRESSION_DATABASE = new UniqueDatabase("logical_server_name", "regression_test")
+            .withDbHistoryPath(DB_HISTORY_PATH);
+        REGRESSION_DATABASE.createAndInitialize();
+
+        config = simpleConfig()
+            .with(MySqlConnectorConfig.SSL_MODE, SecureConnectionMode.REQUIRED)
+            .with(SET_TLS_PROTOCOLS, "TLSv1.7")
+            .build();
+        Filters filters = new Filters.Builder(config).build();
+        context = new MySqlTaskContext(config, filters);
+        context.start();
+        context.source().setBinlogStartPoint("", 0L); // start from beginning
+        context.initializeHistory();
+        reader = new BinlogReader("binlog", context, null);
+
+        // Start reading the binlog ...
+        reader.start();
+    }
+
+    @Test
+    @FixFor( "DBZ-1208" )
+    public void shouldAcceptTls12() {
+        final UniqueDatabase REGRESSION_DATABASE = new UniqueDatabase("logical_server_name", "regression_test")
+            .withDbHistoryPath(DB_HISTORY_PATH);
+        REGRESSION_DATABASE.createAndInitialize();
+
+        config = simpleConfig()
+            .with(MySqlConnectorConfig.SSL_MODE, SecureConnectionMode.REQUIRED)
+            .with(SET_TLS_PROTOCOLS, "TLSv1.2")
+            .build();
+        Filters filters = new Filters.Builder(config).build();
+        context = new MySqlTaskContext(config, filters);
+        context.start();
+        context.source().setBinlogStartPoint("", 0L); // start from beginning
+        context.initializeHistory();
+        reader = new BinlogReader("binlog", context, null);
+
+        // Start reading the binlog ...
+        reader.start();
+        String acceptedTlsVersion = context.getConnectionContext().getSessionVariableForSslVersion();
+        assertEquals("TLSv1.2", acceptedTlsVersion);
     }
 
     private void inconsistentSchema(EventProcessingFailureHandlingMode mode) throws InterruptedException, SQLException {