+++ debezium-connector-postgres/src/test/java/io/debezium/connector/postgresql/RecordsSnapshotProducerIT.java
@@ -13,6 +13,7 @@ import static org.fest.assertions.Assertions.assertThat;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotNull;
 
+import java.util.Arrays;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
@@ -21,9 +22,7 @@ import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicInteger;
 import java.util.stream.Collectors;
 
-import io.debezium.connector.postgresql.snapshot.AlwaysSnapshotter;
-import io.debezium.connector.postgresql.snapshot.InitialOnlySnapshotter;
-import io.debezium.connector.postgresql.spi.Snapshotter;
+import org.apache.kafka.connect.data.Schema;
 import org.apache.kafka.connect.data.Struct;
 import org.apache.kafka.connect.source.SourceRecord;
 import org.junit.After;
@@ -34,6 +33,9 @@ import org.junit.rules.TestRule;
 
 import io.debezium.connector.postgresql.junit.SkipTestDependingOnDatabaseVersionRule;
 import io.debezium.connector.postgresql.junit.SkipWhenDatabaseVersionLessThan;
+import io.debezium.connector.postgresql.snapshot.AlwaysSnapshotter;
+import io.debezium.connector.postgresql.snapshot.InitialOnlySnapshotter;
+import io.debezium.connector.postgresql.spi.Snapshotter;
 import io.debezium.data.Envelope;
 import io.debezium.data.VerifyRecord;
 import io.debezium.doc.FixFor;
@@ -476,4 +478,38 @@ public class RecordsSnapshotProducerIT extends AbstractRecordsProducerTest {
         consumer.process(record -> assertReadRecord(record, expectedValuesByTopicName));
     }
 
+    @Test
+    @FixFor("DBZ-1163")
+    public void shouldGenerateSnapshotForATableWithoutPrimaryKey() throws Exception {
+        final PostgresConnectorConfig config = new PostgresConnectorConfig(
+                TestHelper.defaultConfig()
+                        .with(PostgresConnectorConfig.SNAPSHOT_MODE, PostgresConnectorConfig.SnapshotMode.INITIAL)
+                        .build()
+        );
+        context = new PostgresTaskContext(
+                config,
+                TestHelper.getSchema(config),
+                PostgresTopicSelector.create(config)
+        );
+        snapshotProducer = buildNoStreamProducer(context, config);
+
+        final TestConsumer consumer = testConsumer(1, "public");
+
+        TestHelper.execute("insert into table_without_pk values(1, 1000)");
+
+        //then start the producer and validate all records are there
+        snapshotProducer.start(consumer, e -> {});
+        consumer.await(TestHelper.waitTimeForRecords() * 30, TimeUnit.SECONDS);
+
+        List<SchemaAndValueField> schemaAndValueFields = Arrays.asList(
+                new SchemaAndValueField("id", Schema.INT32_SCHEMA, 1),
+                new SchemaAndValueField("val", Schema.OPTIONAL_INT32_SCHEMA, 1000));
+
+        consumer.process(record -> {
+            assertThat(record.key()).isNull();
+            String actualTopicName = record.topic().replace(TestHelper.TEST_SERVER + ".", "");
+            assertEquals("public.table_without_pk", actualTopicName);
+            assertRecordSchemaAndValues(schemaAndValueFields, record, Envelope.FieldName.AFTER);
+        });
+    }
 }